# unroll jl_typemap_intersection_node_visitor more