this makes the users of typemap-intersection (ambiguity / cache invalidation and ml-lookup) a bit faster by taking better advantage of the tree structure to prune unreachable branches