@JeffBezanson 2016-05-13
Any performance numbers? I want something to drool over :)