@vtjnash 2016-05-13
my test case is https://github.com/JuliaLang/julia/issues/15048, although the benefit is only a few percent (1s out of 30s). it seems to do pretty well on my micro-benchmark however:

```julia
before:
julia> @time using Compat
  0.219453 seconds (89.81 k allocations: 6.619 MB, 27.45% gc time)

after:
julia> @time using Compat
  0.158639 seconds (78.99 k allocations: 6.778 MB)
```

```
before:
julia> @time using Colors
  0.605767 seconds (488.00 k allocations: 22.787 MB, 17.94% gc time)

after:
julia> @time using Colors
  0.506278 seconds (345.27 k allocations: 18.691 MB, 13.26% gc time)
```